requirements.txt
pyyaml
pymupdf
pydantic